requirements.txt
# Europe Job Hunter - Dependencies

# CV Parsing
pdfplumber>=0.10.0

# Scraping
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
